move: #1 ISymbolInfoの場所を移動
