+++ documentationAI/domain/services/package_analyzer/abc.py
@@ -1,30 +1,7 @@
 import abc
 from typing import Dict, Tuple
 
-
-class ISymbolInfo(abc.ABC):
-
-    def __init__(self, namespace: str, symbol_name: str):
-        self.namespace: str = namespace
-        self.symbol_name: str = symbol_name
-
-    @abc.abstractmethod
-    def stringify(self) -> str:
-        pass
-    
-    @classmethod
-    @abc.abstractmethod
-    def parse(cls, stringified: str) -> 'ISymbolInfo':
-        pass
-
-    @abc.abstractmethod
-    def __eq__(self, other: object) -> bool:
-        pass
-
-    @abc.abstractmethod
-    def __str__(self) -> str:
-        pass
-
+from documentationAI.domain.models.document.symbol_info import ISymbolInfo
 
 class IAnalyzerHelper(abc.ABC):
